requirements.txt
Scrapy
# development
ipython
flake8
# test
pytest
pytest-cov
mock
coveralls
vcrpy
requests
